chore: mv buffer2 to buffer
