+++ foundation/buffer.cc
@@ -1,6 +1,6 @@
 /*
  * buffer.cc
- * Copyright (C) 2022 youfa.song <vsyfar@gmail.com>
+ * Copyright (C) 2024 youfa <vsyfar@gmail.com>
  *
  * Distributed under terms of the GPLv2 license.
  */
@@ -16,7 +16,8 @@ namespace ave {
 namespace media {
 
 Buffer::Buffer(size_t capacity)
-    : data_(malloc(capacity)),
+    : buffer_(std::make_unique<base::Buffer>(capacity)),
+      data_(buffer_->data()),
       capacity_(capacity),
       range_offset_(0),
       range_length_(capacity),
@@ -34,19 +35,12 @@ Buffer::Buffer(void* data, size_t capacity)
 // static
 std::shared_ptr<Buffer> Buffer::CreateAsCopy(const void* data,
                                              size_t capacity) {
-  std::shared_ptr<Buffer> res(std::make_shared<Buffer>(capacity));
-  memcpy(res->data(), data, capacity);
-  return res;
+  auto buffer = std::make_shared<Buffer>(capacity);
+  std::memcpy(buffer->data(), data, capacity);
+  return buffer;
 }
 
-Buffer::~Buffer() {
-  if (owns_data_) {
-    if (data_ != nullptr) {
-      free(data_);
-      data_ = nullptr;
-    }
-  }
-}
+Buffer::~Buffer() = default;
 
 void Buffer::setRange(size_t offset, size_t size) {
   AVE_CHECK_LE(offset, capacity_);
@@ -56,7 +50,26 @@ void Buffer::setRange(size_t offset, size_t size) {
   range_length_ = size;
 }
 
-std::shared_ptr<Message> Buffer::meta() {
+void Buffer::ensureCapacity(size_t capacity, bool copy) {
+  if (capacity <= capacity_) {
+    return;
+  }
+
+  if (owns_data_) {
+    auto new_buffer = std::make_unique<base::Buffer>(capacity);
+    if (copy) {
+      std::memcpy(new_buffer->data(), data_, range_length_);
+    }
+    buffer_ = std::move(new_buffer);
+    data_ = buffer_->data();
+  } else {
+    AVE_CHECK(copy);
+    AVE_CHECK(false);
+  }
+  capacity_ = capacity;
+}
+
+std::shared_ptr<Message>& Buffer::meta() {
   if (meta_ == nullptr) {
     meta_ = std::make_shared<Message>();
   }
@@ -64,4 +77,4 @@ std::shared_ptr<Message> Buffer::meta() {
 }
 
 }  // namespace media
-} /* namespace ave */
+}  // namespace ave
