+++ foundation/buffer.h
@@ -1,15 +1,14 @@
 /*
  * buffer.h
- * Copyright (C) 2022 youfa.song <vsyfar@gmail.com>
+ * Copyright (C) 2024 youfa <vsyfar@gmail.com>
  *
  * Distributed under terms of the GPLv2 license.
  */
 
-#ifndef BUFFER_H
-#define BUFFER_H
+#ifndef BUFFER2_H
+#define BUFFER2_H
 
-#include <memory>
-
+#include "base/buffer.h"
 #include "base/constructor_magic.h"
 
 #include "message.h"
@@ -17,33 +16,40 @@
 namespace ave {
 namespace media {
 
+// -----------------------------------------------
+
 class Buffer {
  public:
   Buffer(size_t capacity);
   Buffer(void* data, size_t capacity);
-
-  uint8_t* base() { return static_cast<uint8_t*>(data_); }
-  uint8_t* data() { return static_cast<uint8_t*>(data_) + range_offset_; }
-  size_t capacity() const { return capacity_; }
-  size_t size() const { return range_length_; }
-  size_t offset() const { return range_offset_; }
-  void setRange(size_t offset, size_t size);
+  virtual ~Buffer();
 
   // create buffer from dup of some memory block
   static std::shared_ptr<Buffer> CreateAsCopy(const void* data,
                                               size_t capacity);
+
+  uint8_t* base() { return static_cast<uint8_t*>(data_); }
+  uint8_t* data() { return (static_cast<uint8_t*>(data_) + range_offset_); }
+  size_t capacity() const { return capacity_; }
+  size_t size() const { return range_length_; }
+  size_t offset() const { return range_offset_; }
+  void setRange(size_t offset, size_t size);
+  void ensureCapacity(size_t capacity, bool copy);
+
   void setInt32Data(int32_t data) { int32_data_ = data; }
   int32_t int32Data() const { return int32_data_; }
-  std::shared_ptr<Message> meta();
 
-  virtual ~Buffer();
+  std::shared_ptr<Message>& meta();
 
  private:
   std::shared_ptr<Message> meta_;
+  std::unique_ptr<base::Buffer> buffer_;
+
   void* data_;
   size_t capacity_;
   size_t range_offset_;
   size_t range_length_;
+
   int32_t int32_data_;
   bool owns_data_;
 
@@ -53,4 +59,4 @@ class Buffer {
 }  // namespace media
 } /* namespace ave */
 
-#endif /* !BUFFER_H */
+#endif /* !BUFFER2_H */
